+++ src/modules/edc-demo/components/dashboard-donut-chart/donut-chart-data.ts
@@ -4,4 +4,10 @@ export interface DonutChartData {
   labels: string[];
   datasets: ChartConfiguration<'doughnut'>['data']['datasets'];
   options: ChartConfiguration<'doughnut'>['options'];
+
+  isEmpty: boolean;
+  emptyMessage: string;
+
+  totalLabel: string;
+  totalValue: number;
 }