+++ src/modules/edc-demo/services/contract-negotiation.service.ts
@@ -50,6 +50,10 @@ export class ContractNegotiationService {
     }
   }
 
+  updateNegotiationList() {
+
+  }
+
   isBusy(contractOffer: ContractOffer) {
     return (
       this.runningNegotiations.get(contractOffer.id) !== undefined ||