+++ src/modules/edc-demo/services/transfer-process-utils.ts
@@ -1,0 +1,13 @@
+import {Injectable} from '@angular/core';
+import {TransferProcessDto} from '../../edc-dmgmt-client';
+
+@Injectable({providedIn: 'root'})
+export class TransferProcessUtils {
+  isIncoming(transferProcess: TransferProcessDto): boolean {
+    return transferProcess.dataRequest.connectorId === 'consumer';
+  }
+
+  isOutgoing(transferProcess: TransferProcessDto): boolean {
+    return !this.isIncoming(transferProcess);
+  }
+}